ADD Create round-trip lab orders in testing environments - CircuLab (CHARTS-7068)
The original intent of this work was to allow automated testing of Patient Relinking stories for S69.
These stories require labs order to be sent AND received. A behavior which does not currently exist
in our testing environments.

This feature interceps labs orders and if certain parameters are met, a mock lab result is sent to our
Clinical API which creates an observation message and Review Results task in Web Services.
This task appears in the inbox dashboard of Charts which the provider will review. Currently the
Review Results task will appear with an Unknown Patient status, because this is the use case we are testing in S69.

Web Services has some minor changes to support this feature as well.

Deployment:

As documented in the README, External API now has a dependency to the dotenv gem. A .env file is required at the
root of the project directory. Review the README for the list of required environment variables.

CircuLabs Documentation: http://confluence.carecloud.com/display/CEN/CircuLabs

Related to CHARTS-7032 (https://jira.carecloud.com/browse/CHARTS-7032)
Related to CHARTS-7039 (https://jira.carecloud.com/browse/CHARTS-7039)
Related to CHARTS-7068 (https://jira.carecloud.com/browse/CHARTS-7068)

Signed-off-by: Adrian Martin <amartin@carecloud.com>