PLAT-2114 DocumentsAPI Bugfix: conditionally base64 decode data
The current code assumes files are base64 encoded and decodes
them as such.  As a result, binary files get base64 decoded
and get corrupted and display as blank in the documents app.

As an API, we should not require base64 encoding of files.

Using the Content-Transfer-Encoding header, we can now allow
clients to specify the type of encoding used and act accordingly.

For backwards compatibility, we still assume base64 encoding when
no Content-Transfer-Encoding header is sent.